reformat structure
