+++ semantic_segmentation.py
@@ -1,0 +1,5 @@
+# just a driver for running python code at the root that we need
+from segmentation.train_kitti import train
+
+
+train()
